+++ e2e/runJest.ts
@@ -12,6 +12,7 @@ import dedent from 'dedent';
 import execa = require('execa');
 import * as fs from 'graceful-fs';
 import stripAnsi = require('strip-ansi');
+import {TestPathPatterns} from '@jest/pattern';
 import type {FormattedTestResults} from '@jest/test-result';
 import {normalizeIcons} from '@jest/test-utils';
 import type {Config} from '@jest/types';
@@ -285,5 +286,10 @@ export function getConfig(
     throw error;
   }
 
-  return JSON.parse(stdout);
+  const {testPathPatterns, ...globalConfig} = JSON.parse(stdout);
+
+  return {
+    ...globalConfig,
+    testPathPatterns: new TestPathPatterns(testPathPatterns),
+  };
 }