+++ packages/jest-core/src/__tests__/watch.test.js
@@ -7,6 +7,7 @@
  */
 
 import chalk from 'chalk';
+import {TestPathPatterns} from '@jest/pattern';
 // eslint-disable-next-line import/order
 import {JestHook, KEYS, TestWatcher} from 'jest-watcher';
 
@@ -142,7 +143,7 @@ describe('Watch mode flows', () => {
     pipe = {write: jest.fn()};
     globalConfig = {
       rootDir: '',
-      testPathPatterns: [],
+      testPathPatterns: new TestPathPatterns([]),
       watch: true,
     };
     hasteMapInstances = [{on: () => {}}];
@@ -156,7 +157,7 @@ describe('Watch mode flows', () => {
   });
 
   it('Correctly passing test path pattern', async () => {
-    globalConfig.testPathPatterns = ['test-*'];
+    globalConfig.testPathPatterns = new TestPathPatterns(['test-*']);
 
     await watch(globalConfig, contexts, pipe, hasteMapInstances, stdin);
 
@@ -690,6 +691,14 @@ describe('Watch mode flows', () => {
       ok = ok === '✔︎';
       const pluginPath = `${__dirname}/__fixtures__/plugin_path_config_updater_${option}`;
 
+      const newVal = (() => {
+        if (option === 'testPathPatterns') {
+          return new TestPathPatterns(['a/b', 'c']);
+        }
+
+        return '__JUST_TRYING__';
+      })();
+
       jest.doMock(
         pluginPath,
         () =>
@@ -699,7 +708,7 @@ describe('Watch mode flows', () => {
             }
 
             run(globalConfig, updateConfigAndRun) {
-              updateConfigAndRun({[option]: '__JUST_TRYING__'});
+              updateConfigAndRun({[option]: newVal});
               return Promise.resolve();
             }
           },
@@ -726,7 +735,7 @@ describe('Watch mode flows', () => {
       if (!ok) {
         expector = expector.not;
       }
-      expector.toHaveProperty(option, '__JUST_TRYING__');
+      expector.toHaveProperty(option, newVal);
     },
   );
 
@@ -904,7 +913,7 @@ describe('Watch mode flows', () => {
     await nextTick();
 
     expect(runJestMock.mock.calls[0][0].globalConfig).toMatchObject({
-      testPathPatterns: ['file'],
+      testPathPatterns: {patterns: ['file']},
       watch: true,
       watchAll: false,
     });
@@ -928,7 +937,7 @@ describe('Watch mode flows', () => {
 
     expect(runJestMock.mock.calls[1][0].globalConfig).toMatchObject({
       testNamePattern: 'test',
-      testPathPatterns: ['file'],
+      testPathPatterns: {patterns: ['file']},
       watch: true,
       watchAll: false,
     });