+++ packages/jest-pattern/src/index.ts
@@ -5,7 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-module.exports = function (globalConfig, projectConfig) {
-  console.log(globalConfig.testPathPatterns.patterns);
-  console.log(projectConfig.cache);
-};
+export {
+  TestPathPatterns,
+  TestPathPatternsExecutor,
+  type TestPathPatternsExecutorOptions,
+} from './TestPathPatterns';