+++ packages/jest-reporters/src/SummaryReporter.ts
@@ -12,7 +12,6 @@ import type {
   TestContext,
 } from '@jest/test-result';
 import type {Config} from '@jest/types';
-import {TestPathPatterns} from 'jest-util';
 import BaseReporter from './BaseReporter';
 import getResultHeader from './getResultHeader';
 import getSnapshotSummary from './getSnapshotSummary';
@@ -212,7 +211,7 @@ export default class SummaryReporter extends BaseReporter {
     testContexts: Set<TestContext>,
     globalConfig: Config.GlobalConfig,
   ) {
-    const testPathPatterns = TestPathPatterns.fromGlobalConfig(globalConfig);
+    const testPathPatterns = globalConfig.testPathPatterns;
 
     const getMatchingTestsInfo = () => {
       const prefix = globalConfig.findRelatedTests